+++ src/renderer/src/components/project/TodoWindow.tsx
@@ -1,5 +1,5 @@
-import { useState, useEffect } from 'react';
-import { MdExpandLess, MdOutlineChecklist, MdAdd } from 'react-icons/md';
+import { useEffect, useState } from 'react';
+import { MdAdd, MdExpandLess, MdOutlineChecklist } from 'react-icons/md';
 import { useTranslation } from 'react-i18next';
 import { TodoItem } from '@common/types';